backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
openai
pydantic
python-dotenv
uvicorn[standard]
google-genai
sqlalchemy
aiosqlite
httpx
chromadb